requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
azure-core==1.37.0
botbuilder-core==4.17.0
botbuilder-schema==4.17.0
botframework-connector==4.17.0
botframework-streaming==4.17.0
certifi==2025.11.12
cffi==2.0.0; python_full_version >= "3.9" and platform_python_implementation != "PyPy"
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6; sys_platform == "win32" or platform_system == "Windows"
cryptography==46.0.3
fastapi==0.124.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
isodate==0.7.2
jsonpickle==1.4.2
msal==1.34.0
msrest==0.7.1
multidict==6.7.0
oauthlib==3.3.1
propcache==0.4.1
pycparser==2.23; python_full_version >= "3.9" and platform_python_implementation != "PyPy" and implementation_name != "PyPy"
pydantic==2.12.5
pydantic-core==2.41.5
pyjwt[crypto]==2.10.1
python-dotenv==1.2.1
pyyaml==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
starlette==0.50.0
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.2
uvicorn[standard]==0.38.0
uvloop==0.22.1; (sys_platform != "cygwin" and sys_platform != "win32") and platform_python_implementation != "PyPy"
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
